+++ common/db/src/main/java/is/codion/common/db/operation/DefaultFunctionType.java
@@ -23,7 +23,7 @@ import java.io.Serializable;
 
 import static java.util.Objects.requireNonNull;
 
-final class DefaultFunctionType<C, T, R> implements FunctionType<C, T, R>, Serializable {
+final class DefaultFunctionType<C, P, R> implements FunctionType<C, P, R>, Serializable {
 
 	@Serial
 	private static final long serialVersionUID = 1;