+++ common/db/src/main/java/is/codion/common/db/operation/FunctionType.java
@@ -20,10 +20,10 @@ package is.codion.common.db.operation;
 
 /**
  * @param <C> the connection type
- * @param <T> the function parameter type
+ * @param <P> the function parameter type
  * @param <R> the function result type
  */
-public interface FunctionType<C, T, R> {
+public interface FunctionType<C, P, R> {
 
 	/**
 	 * @return the function name
@@ -34,11 +34,11 @@ public interface FunctionType<C, T, R> {
 	 * Creates a {@link FunctionType} with the given name and types.
 	 * @param name the name
 	 * @param <C> the connection type
-	 * @param <T> the function parameter type
+	 * @param <P> the function parameter type
 	 * @param <R> the function result type
 	 * @return a new {@link FunctionType}
 	 */
-	static <C, T, R> FunctionType<C, T, R> functionType(String name) {
+	static <C, P, R> FunctionType<C, P, R> functionType(String name) {
 		return new DefaultFunctionType<>(name);
 	}
 }