+++ framework/server/src/main/java/is/codion/framework/server/DefaultRemoteEntityConnection.java
@@ -102,31 +102,31 @@ final class DefaultRemoteEntityConnection extends AbstractRemoteEntityConnection
 	}
 
 	@Override
-	public <T, R, P> R report(ReportType<T, R, P> reportType, P parameter) {
+	public <T, P, R> R report(ReportType<T, P, R> reportType, P parameter) {
 		synchronized (connectionProxy) {
 			return connectionProxy.report(reportType, parameter);
 		}
 	}
 
 	@Override
-	public <C extends EntityConnection, T> void execute(ProcedureType<C, T> procedureType) {
+	public <C extends EntityConnection, P> void execute(ProcedureType<C, P> procedureType) {
 		execute(procedureType, null);
 	}
 
 	@Override
-	public <C extends EntityConnection, T> void execute(ProcedureType<C, T> procedureType, T parameter) {
+	public <C extends EntityConnection, P> void execute(ProcedureType<C, P> procedureType, P parameter) {
 		synchronized (connectionProxy) {
 			connectionProxy.execute(procedureType, parameter);
 		}
 	}
 
 	@Override
-	public <C extends EntityConnection, T, R> R execute(FunctionType<C, T, R> functionType) {
+	public <C extends EntityConnection, P, R> R execute(FunctionType<C, P, R> functionType) {
 		return execute(functionType, null);
 	}
 
 	@Override
-	public <C extends EntityConnection, T, R> R execute(FunctionType<C, T, R> functionType, T parameter) {
+	public <C extends EntityConnection, P, R> R execute(FunctionType<C, P, R> functionType, P parameter) {
 		synchronized (connectionProxy) {
 			return connectionProxy.execute(functionType, parameter);
 		}