+++ framework/servlet/src/main/java/is/codion/framework/servlet/EntityService.java
@@ -559,7 +559,7 @@ public final class EntityService implements AuxiliaryServer {
 			try {
 				RemoteEntityConnection connection = authenticate(context);
 				List<Object> parameters = deserialize(context.req());
-				ReportType<?, ?, Object> reportType = (ReportType<?, ?, Object>) parameters.get(0);
+				ReportType<?, Object, ?> reportType = (ReportType<?, Object, ?>) parameters.get(0);
 				context.status(HttpStatus.OK_200)
 								.contentType(ContentType.APPLICATION_OCTET_STREAM)
 								.result(serialize(connection.report(reportType, parameters.get(1))));